Move collections to a new module
